requirements.txt
dicttoxml==1.7.4
lxml==4.4.2
pymongo==3.10.1
PyYAML==5.3
xmltodict==0.12.0
